requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pillow==10.2.0
beanie==1.26.0
email-validator==2.1.0.post1

# Testing dependencies
pytest==8.4.1
pytest-asyncio==1.1.0
httpx==0.26.0
pytest-mock==3.12.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0